Issue #2853988 : Optimized the post Drupal Scaffold Procedure

+++ src/composer/ScriptHandler.php
@@ -7,7 +7,7 @@
 
 namespace Varbase\composer;
 
-use Composer\EventDispatcher\Event;
+use Symfony\Component\Filesystem\Filesystem;
 
 class ScriptHandler {
 
@@ -27,19 +27,29 @@ class ScriptHandler {
    * @param \Composer\EventDispatcher\Event $event
    *   The script event.
    */
-  public static function postDrupalScaffoldProcedure(Event $event) {
-    //Create staging robots file
-    copy(getcwd() . '/src/assets/robots-staging.txt', static::getDrupalRoot(getcwd()) . '/robots-staging.txt');
-    //Alter .htaccess file
-    $htaccess_path = static::getDrupalRoot(getcwd()) . '/.htaccess';
-    $htaccess_lines = file($htaccess_path);
-    $lines = [];
-    foreach ($htaccess_lines as $line) {
-      $lines[] = $line;
-      if(strpos($line, "RewriteEngine on") !== FALSE) {
-        $lines = array_merge($lines, file(getcwd() . '/src/assets/htaccess_extra'));
-      }
+  public static function postDrupalScaffoldProcedure(\Composer\EventDispatcher\Event $event) {
+
+    $fs = new Filesystem();
+    $root = static::getDrupalRoot(getcwd());
+
+    if ($fs->exists($root . '/profiles/varbase/src/assets/robots-staging.txt')) {
+      //Create staging robots file
+      copy($root . '/profiles/varbase/src/assets/robots-staging.txt', $root . '/robots-staging.txt');
     }
-    file_put_contents($htaccess_path, $lines);
+
+    if ($fs->exists($root . '/.htaccess')) {
+      //Alter .htaccess file
+      $htaccess_path = $root . '/.htaccess';
+      $htaccess_lines = file($htaccess_path);
+      $lines = [];
+      foreach ($htaccess_lines as $line) {
+        $lines[] = $line;
+        if (strpos($line, "RewriteEngine on") !== FALSE
+          && $fs->exists($root . '/profiles/varbase/src/assets/htaccess_extra')) {
+          $lines = array_merge($lines, file($root . '/profiles/varbase/src/assets/htaccess_extra'));
+        }
+      }
+      file_put_contents($htaccess_path, $lines);
+    }
   }
 }
